requirements.txt
towhee
gradio
opencv-python
opencv-python-headless
jupyter
pymilvus
langchain
langchain_openai
langchain_core
pandas
torch